# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "PSP System App",
        "contact": {
            "name": "Mahmoud Mohamed",
            "url": "https://www.linkedin.com/in/mahmoud-mohamed-955932b5/",
            "email": "m.mohamed@cat.com.eg"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/account/{guard}/me": {
            "get": {
                "tags": [
                    "account"
                ],
                "summary": "Get the current logged in user",
                "parameters": [
                    {
                        "name": "guard",
                        "in": "path",
                        "description": "userApi , companyApi , pharmacyAdminApi , pharmacyUsersApi ",
                        "required": true,
                        "type": "string",
                        "format": "string",
                        "default": "userApi"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "object"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/account/{guard}/my-product": {
            "get": {
                "tags": [
                    "account"
                ],
                "summary": "Get my product",
                "parameters": [
                    {
                        "name": "guard",
                        "in": "path",
                        "description": "userApi , companyApi , pharmacyAdminApi , pharmacyUsersApi ",
                        "required": true,
                        "type": "string",
                        "format": "string",
                        "default": "userApi"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "object"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/account/{guard}/my-patients-cards": {
            "get": {
                "tags": [
                    "account (doctor)"
                ],
                "summary": "Get my product",
                "parameters": [
                    {
                        "name": "guard",
                        "in": "path",
                        "description": "userApi , companyApi , pharmacyAdminApi , pharmacyUsersApi ",
                        "required": true,
                        "type": "string",
                        "format": "string",
                        "default": "userApi"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "object"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/account/{guard}/update": {
            "post": {
                "tags": [
                    "account"
                ],
                "summary": "update My Account",
                "parameters": [
                    {
                        "name": "guard",
                        "in": "path",
                        "description": "userApi , companyApi , pharmacyAdminApi , pharmacyUsersApi ",
                        "required": true,
                        "type": "string",
                        "format": "string",
                        "default": "userApi"
                    },
                    {
                        "name": "name",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string",
                        "default": "mahmoud"
                    },
                    {
                        "name": "email",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string",
                        "default": "rep1@gmail.com"
                    },
                    {
                        "name": "photo",
                        "in": "formData",
                        "type": "file"
                    },
                    {
                        "name": "national_id",
                        "in": "formData",
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "age",
                        "in": "formData",
                        "type": "integer"
                    },
                    {
                        "name": "sex",
                        "in": "formData",
                        "description": "1=>male , 2=>Female",
                        "type": "integer"
                    },
                    {
                        "name": "address",
                        "in": "formData",
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "phone",
                        "in": "formData",
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "phone2",
                        "in": "formData",
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "diagnosis",
                        "in": "formData",
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "specialty",
                        "in": "formData",
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "preferred_distributor",
                        "in": "formData",
                        "type": "string",
                        "format": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "token"
                    },
                    "400": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/account/{guard}/update-password": {
            "post": {
                "tags": [
                    "account"
                ],
                "summary": "update My Password",
                "parameters": [
                    {
                        "name": "guard",
                        "in": "path",
                        "description": "userApi , companyApi , pharmacyAdminApi , pharmacyUsersApi ",
                        "required": true,
                        "type": "string",
                        "format": "string",
                        "default": "userApi"
                    },
                    {
                        "name": "current_password",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User Model"
                    },
                    "400": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/account/{guard}/my-products": {
            "get": {
                "tags": [
                    "account"
                ],
                "summary": "get my account products",
                "responses": {
                    "200": {
                        "description": "object"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        }
    },
...
    "securityDefinitions": {
        "jwt": {
            "type": "apiKey",
            "description": "Value: Bearer \\<token\\><br><br>",
            "name": "Authorization",
            "in": "header"
        }
    }
}